# ci-build: move tests.out to target so it wont cause rebuild