It saves about 2m46s on CircleCI.

Previous: 

```
#!/bin/bash -eo pipefail
make trace_test
env CI=true SKIP_FORMAT_CHECK=true FAIL_POINT=1 /home/circleci/tikv-ci/ci-build/test.sh
make[1]: Entering directory `/home/circleci/tikv-ci'
    Finished dev [unoptimized + debuginfo] target(s) in 0.39s
make[1]: Leaving directory `/home/circleci/tikv-ci'
make[1]: Entering directory `/home/circleci/tikv-ci'
# When SIP is enabled, DYLD_LIBRARY_PATH will not work in subshell, so we have to set it
# again here. LOCAL_DIR is defined in .travis.yml.
export DYLD_LIBRARY_PATH=":/home/circleci/.local/lib:/home/circleci/.local/lib:/home/circleci/.local/lib:/lib" && \
	export LOG_LEVEL=DEBUG && \
	export RUST_BACKTRACE=1 && \
	cargo test --features "default portable sse"  -- --nocapture && \
	cargo test --features "default portable sse" --bench benches  -- --nocapture  && \
	if [[ "`uname`" == "Linux" ]]; then \
		export MALLOC_CONF=prof:true,prof_active:false && \
		cargo test --features "default portable sse"  --bin tikv-server -- --nocapture --ignored; \
	fi
 INFO 2018-07-23T03:35:24Z: cargo::core::compiler::fingerprint: fingerprint error for tikv v2.1.0-beta (file:///home/circleci/tikv-ci): precalculated components have changed: 1532316923.853428612s (/home/circleci/tikv-ci/tests.out) != 1532316132.273408917s (/home/circleci/tikv-ci/.circleci/config.yml)
   Compiling tikv v2.1.0-beta (file:///home/circleci/tikv-ci)
    Finished dev [unoptimized + debuginfo] target(s) in 2m 46s
     Running target/debug/deps/tikv-ad2709b06283c708

running 506 tests
```

Now:

```
#!/bin/bash -eo pipefail
make trace_test
env CI=true SKIP_FORMAT_CHECK=true FAIL_POINT=1 /home/circleci/tikv-ci/ci-build/test.sh
make[1]: Entering directory `/home/circleci/tikv-ci'
    Finished dev [unoptimized + debuginfo] target(s) in 0.37s
make[1]: Leaving directory `/home/circleci/tikv-ci'
make[1]: Entering directory `/home/circleci/tikv-ci'
# When SIP is enabled, DYLD_LIBRARY_PATH will not work in subshell, so we have to set it
# again here. LOCAL_DIR is defined in .travis.yml.
export DYLD_LIBRARY_PATH=":/home/circleci/.local/lib:/home/circleci/.local/lib:/home/circleci/.local/lib:/lib" && \
	export LOG_LEVEL=DEBUG && \
	export RUST_BACKTRACE=1 && \
	cargo test --features "default portable sse"  -- --nocapture && \
	cargo test --features "default portable sse" --bench benches  -- --nocapture  && \
	if [[ "`uname`" == "Linux" ]]; then \
		export MALLOC_CONF=prof:true,prof_active:false && \
		cargo test --features "default portable sse"  --bin tikv-server -- --nocapture --ignored; \
	fi
    Finished dev [unoptimized + debuginfo] target(s) in 0.24s  <---- Here
     Running target/debug/deps/tikv-ad2709b06283c708

running 506 tests
```