@BusyJay 2018-07-23
I think what you need to change is `build.rs`. Otherwise, any changes in the directory will trigger a rebuild.