@BusyJay 2018-07-25
So build time will not change any more?