@overvenus 2018-07-26
@breeswish To avoid re-run if any file inside the crate root changes.
https://doc.rust-lang.org/cargo/reference/build-scripts.html#outputs-of-the-build-script

@BusyJay It will change as long as we commit.