@breeswish 2018-07-26
How about rerun if only content in directories like `src/`, `benches/` is changed?