@overvenus 2018-07-27
@breeswish It is not enough. It wont re-run even if source code changes.

```
$ ls -a --full-time | grep src
drwxr-xr-x 10 neil neil  4096 2018-07-27 10:45:14.401240677 +0800 src
$ touch src/coprocessor/mod.rs
$ ls -a --full-time | grep src
drwxr-xr-x 10 neil neil  4096 2018-07-27 10:45:14.401240677 +0800 src
```