@breeswish 2018-07-27
@overvenus As [the documentation says](https://doc.rust-lang.org/cargo/reference/build-scripts.html), you also need a rule for everything inside the directory. You can have a try.

> If this path points to a directory the entire directory will not be traversed for changes -- only changes to the timestamp of the directory itself (which corresponds to some types of changes within the directory, depending on platform) will trigger a rebuild. To request a re-run on any changes within an entire directory, print a line for the directory and another line for everything inside it, recursively.