@overvenus 2018-07-27
Actually, we dont need to traverse `src`. Build.rs generates commit hash and branch name, and it does not care about `src`.